feat: 퀴즈 캐러셀 적용
