+++ src/app/learn/reading/detail/[id]/page.tsx
@@ -10,6 +10,7 @@ import { useState } from 'react';
 import ResponsiveSidebar from '@/components/ResponsiveSidebar';
 import Quiz from '@/components/quiz/Quiz';
 import quizData from '@/mock/quizData.json';
+import QuizCarousel from '@/components/quiz/QuizCarousel';
 
 export default function DetailReadingPage() {
   const [showTranslate, toggleShowTranlation] = useState(true);
@@ -78,7 +79,7 @@ export default function DetailReadingPage() {
               ))}
             </ul>
           </div>
-          <Quiz quizData={quizData[0]} />
+          <QuizCarousel quizListData={quizData} />
         </div>
 
         <div className="fixed right-[60px] bottom-[76px] md:bottom-[16px]">
